Add a test to determine whether or not the game is over
This is functioning, but in a very limited way. One of the biggest issues is with events. Currently a tile will keep responding to clicks even though a response is queued for the initial click (you can keep clicking on a tile rapidly, and that will register a match event). Basically it's broken.